+++ src/console/controllers/ImportController.php
@@ -44,14 +44,10 @@ class ImportController extends Controller
     /**
      * Import a json file structure.
      *
-     * @return mixed
+     * @param string $filename
      */
-    public function actionIndex()
+    public function actionIndex($filename)
     {
-        $result = 'something';
-
         echo "Welcome to the console ImportController actionIndex() method\n";
-
-        return $result;
     }
 }