+++ CHANGELOG.md
@@ -6,6 +6,11 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+## [2.2.7] - 2018-04-02
+### Fixed
+- Asset field source mapping on import/export.
+- Global field layout not importing.
+
 ## [2.2.6] - 2018-03-22
 ### Fixed
 - SuperTable import source mapping.
@@ -129,7 +134,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - Field Group Importing
 - Field Importing
 
-[Unreleased]: https://github.com/pennebaker/craft-architect/compare/2.2.6...develop
+[Unreleased]: https://github.com/pennebaker/craft-architect/compare/2.2.7...develop
+[2.2.7]: https://github.com/pennebaker/craft-architect/compare/2.2.6...2.2.7
 [2.2.6]: https://github.com/pennebaker/craft-architect/compare/2.2.5...2.2.6
 [2.2.5]: https://github.com/pennebaker/craft-architect/compare/2.2.4...2.2.5
 [2.2.4]: https://github.com/pennebaker/craft-architect/compare/2.2.3...2.2.4